requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.115.12
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
passlib==1.7.4
proto-plus==1.26.1
protobuf==5.29.4
pyasn1
pyasn1_modules
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
